refactor(#8): msw 로그 api 분리, 시간 형식 절대시각으로 변경
