+++ src/mocks/handlers/handlers.ts
@@ -1,9 +1,11 @@
 import { authHandlers } from "./auth-handlers";
 import { dashboardHandlers } from "./dashboard-handlers";
 import { inventoryHandlers } from "./inventory-handlers";
+import { dungeonLogHandlers } from "./dungeon-log-handlers";
 
 export const handlers = [
   ...authHandlers,
   ...dashboardHandlers,
+  ...dungeonLogHandlers,
   ...inventoryHandlers,
 ];
